pyproject.toml
[project]
name = "cibmtr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Felix Neumann", email = "felix.maximilian.neumann@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "lifelines>=0.30.0",
    "matplotlib>=3.10.0",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "scikit-learn>=1.6.1",
    "tqdm>=4.67.1",
    "wandb>=0.19.6",
]

[project.scripts]
cibmtr = "cibmtr:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
local = [
    "torch>=2.6.0",
]
